Added TFTest and moved some stuff
