+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/old/VisionDepotSide.java
@@ -1,6 +1,8 @@
-package org.firstinspires.ftc.teamcode.autonomous;
+package org.firstinspires.ftc.teamcode.autonomous.old;
 
+import com.qualcomm.hardware.rev.RevBlinkinLedDriver;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.ClassFactory;
@@ -8,13 +10,14 @@ import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
 import org.firstinspires.ftc.robotcore.external.tfod.Recognition;
 import org.firstinspires.ftc.robotcore.external.tfod.TFObjectDetector;
 import org.firstinspires.ftc.teamcode.HardwareDefinitions;
+import org.firstinspires.ftc.teamcode.TFLiteHandler;
+import org.firstinspires.ftc.teamcode.autonomous.AutonomousDefinitions;
 
 import java.util.ArrayList;
 import java.util.List;
 
-@Autonomous(name = "VisionCraterSide - No Landing", group = "Vision")
-public class VisionCraterSideNoLand extends AutonomousDefinitions {
-
+@Autonomous(name = "VisionDepotSide - With Landing", group = "Vision")
+public class VisionDepotSide extends AutonomousDefinitions {
 
     private static final String TFOD_MODEL_ASSET = "RoverRuckus.tflite";
     private static final String LABEL_GOLD_MINERAL = "Gold Mineral";
@@ -36,88 +39,109 @@ public class VisionCraterSideNoLand extends AutonomousDefinitions {
     public void runOpMode() {
 
         init(hardwareMap);
+
         //initIMU(hardwareMap);
 
         initTFodAndVuforia();
-
 
         telemetry.addData("Robot is initialized", "");
         telemetry.update();
         waitForStart();
 
-
+        autonLEDPattern = RevBlinkinLedDriver.BlinkinPattern.HOT_PINK;
+        LEDController.setPattern(autonLEDPattern);
 
         //add movement to
 
         markerDropperOuter.setPosition(markerDropperOuterHold);
 
-        //dropFromLander();
+        dropFromLander(false);
         encoderDrive(0.4 ,14, 14, 5);
         //moveLanderWithEncoder((38*4), 8);
         encoderTurn(0.25, 105, false, 5);
-        encoderDrive(0.4, 7, 7, 5);
+        encoderDrive(0.4, 5, 5, 5);
 
         markerDropperOuter.setPosition(markerDropperOuterRelease);
 
-        detectGold_inferRight(5);
+        detectGold_inferRight(2);
 
         //movement stuff
 
         markerDropperOuter.setPosition(markerDropperOuterHold);
 
 
-        switch(getMineralPosition(true)){
-
+        switch(getMineralPosition(false)){
             case 1:
-
-                //encoderDrive(0.4, 3, 3, 5);
+                //encoderDrive(0.4, 7, 7, 5);
                 encoderTurn(0.25, 70, true, 5); //turn left and drive towards the gold
-                encoderDrive(0.7, 26, 26, 10);
-                moveLeftTread(0.4, 15, 8);
-                //encoderDrive(0.35, -12, -12, 10);
-                //encoderTurn(0.25, 105, false, 5);
+                encoderDrive(0.4, 27, 27, 10);
+                encoderTurn(0.25, 70, true, 5);
+                encoderDrive(0.4, 25, 21, 5);
+                encoderTurn(0.25, 130, false, 5);
+
+                //drop the marker
+                moveBoxMechanism(2, 2);
+                dropMarker();
+                //moveBoxMechanism(-2, 3);
+
+
+                //encoderTurn(0.25, 100, false, 5);
 
                 break;
 
             case 2:
 
-                encoderDrive(0.35, -5, -5, 5); //drive straight towards the gold
+                encoderDrive(0.35, -5.5, -5.5, 5); //drive straight towards the gold
                 encoderTurn(0.25, 105, true, 5);
-                encoderDrive(0.7, 26, 26, 10);
-                //encoderDrive(0.35, -15, -15, 10);
-                //encoderTurn(0.25, 105, false, 5);
-                //encoderDrive(0.4, 14, 14, 10);
+                encoderDrive(0.35, 46, 46, 10);
+                encoderTurn(0.25, 105, false, 5);
+
+                //drop the marker
+                moveBoxMechanism(2, 2);
+                dropMarker();
+                //moveBoxMechanism(-2, 3);
+
+                encoderDrive(0.4, 10, 10, 5);
 
                 break;
 
+
             case 3:
 
-                encoderDrive(0.4, -20, -20, 5);
-                encoderTurn(0.25, 115, true, 5); //turn right and drive towards the gold
-                encoderDrive(0.7, 26, 26, 10);
-                moveRightTread(0.4, 10, 5);
-                //encoderDrive(0.35, -15, -15, 10);
-                //encoderTurn(0.25, 105, false, 5);
-                //encoderDrive(0.4, 29, 29, 10);
+                encoderDrive(0.4, -17, -17, 5);
+                encoderTurn(0.25, 130, true, 5); //turn right and drive towards the gold
+                encoderDrive(0.35, 28, 28, 10);
+                encoderTurn(0.25, 75, false, 5);
+
+                encoderDrive(0.4, 29, 29, 5);
+                encoderTurn(0.25, 80, false, 5);
+
+                //drop the marker
+                moveBoxMechanism(2, 2);
+                dropMarker();
+                //moveBoxMechanism(-2, 3);
+
+                encoderDrive(0.4, 10, 10, 5);
 
                 break;
 
             default:
 
-                //encoderDrive(0.4, 3, 3, 5);
-                encoderTurn(0.25, 70, true, 5); //turn left and drive towards the gold
-                encoderDrive(0.7, 26, 26, 10);
-                moveLeftTread(0.4, 15, 8);
-                //encoderDrive(0.35, -12, -12, 10);
-                //encoderTurn(0.25, 105, false, 5);
+                encoderDrive(0.4, 7, 7, 5);
+                encoderTurn(0.25, 90, true, 5); //turn left and drive towards the gold
+                encoderDrive(0.35, 23, 23, 10);
+                encoderTurn(0.25, 75, true, 5);
+                encoderDrive(0.4, 21, 21, 5);
+                encoderTurn(0.25, 105, false, 5);
+
+                //drop the marker
+                moveBoxMechanism(2, 2);
+                dropMarker();
+                //moveBoxMechanism(-2, 3);
 
                 break;
         }
 
-        //finesse the robot over the crater boundary
-        intakeMotor.setPower(1);
-        sleep(3000);
-        intakeMotor.setPower(0);
 
 
     }
@@ -209,7 +233,7 @@ public class VisionCraterSideNoLand extends AutonomousDefinitions {
                     }
                 }
 
-                if(timer.seconds() >= 2){ //give the robot a little bit of time to come to a stop before recording values
+                if(timer.seconds() >= 0.5){ //give the robot a little bit of time to come to a stop before recording values
                     detectionValues.add(currentDetectionValue);
                 }
             }
@@ -249,6 +273,90 @@ public class VisionCraterSideNoLand extends AutonomousDefinitions {
 
             }
 
+        }
+    }
+
+    /*
+    Drive forwards until two minerals are centered in the camera's view
+    Speed should be a low value so that the robot doesn't overshoot the minerals
+     */
+    public void driveToMinerals(double maxTimeS){
+
+        int CAMERA_MAX_LEFT = 0;
+        int CAMERA_MAX_RIGHT = 1920; //this should be the same as the horizontal number of pixels of the camera
+        int ERROR_THRESHOLD = 20; //maximum tolerance for the minerals being off center
+
+        boolean centered = false;
+
+        int mineral1XLeft;
+        int mineral2XRight;
+
+        int leftDifference;
+        int rightDifference;
+
+        timer.reset();
+
+        if (opModeIsActive()) {
+            /** Start Tensor Flow Object Detection. */
+            if (tfod != null) {
+                tfod.activate();
+            }
+
+            motorL1.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            motorL2.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            motorR1.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            motorR2.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+
+            motorL1.setPower(0.3);
+            motorL2.setPower(0.3);
+            motorR1.setPower(0.3);
+            motorR2.setPower(0.3);
+
+            while (opModeIsActive() && timer.seconds() < maxTimeS && !centered) {
+                if (tfod != null) {
+                    List<Recognition> updatedRecognitions = tfod.getUpdatedRecognitions();
+                    if (updatedRecognitions != null) {
+                        telemetry.addData("# Object Detected", updatedRecognitions.size());
+                        if (updatedRecognitions.size() == 2) {
+                            motorL1.setPower(0.1);
+                            motorL2.setPower(0.1);
+                            motorR1.setPower(0.1);
+                            motorR2.setPower(0.1);
+
+                            mineral1XLeft = (int) updatedRecognitions.get(0).getLeft();
+                            mineral2XRight = (int) updatedRecognitions.get(1).getRight();
+
+                            leftDifference = Math.abs(CAMERA_MAX_LEFT - mineral1XLeft);
+                            rightDifference = Math.abs(CAMERA_MAX_RIGHT - mineral2XRight);
+
+                            if(leftDifference > rightDifference &&
+                                    Math.abs(leftDifference - rightDifference) > ERROR_THRESHOLD){
+
+                                motorL1.setPower(0.1); //the minerals are too far right in the camera frame, so keep driving
+                                motorL2.setPower(0.1);
+                                motorR1.setPower(0.1);
+                                motorR2.setPower(0.1);
+                            } else if(leftDifference < rightDifference &&
+                                    Math.abs(leftDifference - rightDifference) > ERROR_THRESHOLD){
+
+                                motorL1.setPower(-0.1); //the minerals are too far left in the camera frame, so drive in reverse
+                                motorL2.setPower(-0.1);
+                                motorR1.setPower(-0.1);
+                                motorR2.setPower(-0.1);
+                            } else {
+                                motorL1.setPower(0); //the minerals are within the threshold, so stop movement
+                                motorL2.setPower(0);
+                                motorR1.setPower(0);
+                                motorR2.setPower(0);
+
+                                centered = true; //breaks loop
+
+                            }
+
+                        }
+                    }
+                }
+            }
         }
     }
