+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/old/VisionDepotSideNoLand.java
@@ -1,8 +1,6 @@
-package org.firstinspires.ftc.teamcode.autonomous;
+package org.firstinspires.ftc.teamcode.autonomous.old;
 
-import com.qualcomm.hardware.rev.RevBlinkinLedDriver;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.ClassFactory;
@@ -10,13 +8,13 @@ import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
 import org.firstinspires.ftc.robotcore.external.tfod.Recognition;
 import org.firstinspires.ftc.robotcore.external.tfod.TFObjectDetector;
 import org.firstinspires.ftc.teamcode.HardwareDefinitions;
-import org.firstinspires.ftc.teamcode.TFLiteHandler;
+import org.firstinspires.ftc.teamcode.autonomous.AutonomousDefinitions;
 
 import java.util.ArrayList;
 import java.util.List;
 
-@Autonomous(name = "VisionDepotSide - With Landing", group = "Vision")
-public class VisionDepotSide extends AutonomousDefinitions {
+@Autonomous(name = "VisionDepotSide - No Landing", group = "Vision")
+public class VisionDepotSideNoLand extends AutonomousDefinitions {
 
     private static final String TFOD_MODEL_ASSET = "RoverRuckus.tflite";
     private static final String LABEL_GOLD_MINERAL = "Gold Mineral";
@@ -38,7 +36,6 @@ public class VisionDepotSide extends AutonomousDefinitions {
     public void runOpMode() {
 
         init(hardwareMap);
-
         //initIMU(hardwareMap);
 
         initTFodAndVuforia();
@@ -47,29 +44,28 @@ public class VisionDepotSide extends AutonomousDefinitions {
         telemetry.update();
         waitForStart();
 
-        autonLEDPattern = RevBlinkinLedDriver.BlinkinPattern.HOT_PINK;
-        LEDController.setPattern(autonLEDPattern);
+
 
         //add movement to
 
         markerDropperOuter.setPosition(markerDropperOuterHold);
 
-        dropFromLander(false);
+        //dropFromLander();
         encoderDrive(0.4 ,14, 14, 5);
         //moveLanderWithEncoder((38*4), 8);
         encoderTurn(0.25, 105, false, 5);
-        encoderDrive(0.4, 5, 5, 5);
+        encoderDrive(0.4, 7, 7, 5);
 
         markerDropperOuter.setPosition(markerDropperOuterRelease);
 
-        detectGold_inferRight(2);
+        detectGold_inferRight(5);
 
         //movement stuff
 
         markerDropperOuter.setPosition(markerDropperOuterHold);
 
 
-        switch(getMineralPosition(false)){
+        switch(getMineralPosition(true)){
             case 1:
                 //encoderDrive(0.4, 7, 7, 5);
                 encoderTurn(0.25, 70, true, 5); //turn left and drive towards the gold
@@ -81,7 +77,7 @@ public class VisionDepotSide extends AutonomousDefinitions {
                 //drop the marker
                 moveBoxMechanism(2, 2);
                 dropMarker();
-                //moveBoxMechanism(-2, 3);
+                moveBoxMechanism(-2, 2);
 
 
                 //encoderTurn(0.25, 100, false, 5);
@@ -98,7 +94,7 @@ public class VisionDepotSide extends AutonomousDefinitions {
                 //drop the marker
                 moveBoxMechanism(2, 2);
                 dropMarker();
-                //moveBoxMechanism(-2, 3);
+                moveBoxMechanism(-2, 1);
 
                 encoderDrive(0.4, 10, 10, 5);
 
@@ -118,7 +114,7 @@ public class VisionDepotSide extends AutonomousDefinitions {
                 //drop the marker
                 moveBoxMechanism(2, 2);
                 dropMarker();
-                //moveBoxMechanism(-2, 3);
+                moveBoxMechanism(-2, 1);
 
                 encoderDrive(0.4, 10, 10, 5);
 
@@ -136,7 +132,7 @@ public class VisionDepotSide extends AutonomousDefinitions {
                 //drop the marker
                 moveBoxMechanism(2, 2);
                 dropMarker();
-                //moveBoxMechanism(-2, 3);
+                moveBoxMechanism(-2, 1);
 
                 break;
         }
@@ -232,7 +228,7 @@ public class VisionDepotSide extends AutonomousDefinitions {
                     }
                 }
 
-                if(timer.seconds() >= 0.5){ //give the robot a little bit of time to come to a stop before recording values
+                if(timer.seconds() >= 2){ //give the robot a little bit of time to come to a stop before recording values
                     detectionValues.add(currentDetectionValue);
                 }
             }
@@ -272,90 +268,6 @@ public class VisionDepotSide extends AutonomousDefinitions {
 
             }
 
-        }
-    }
-
-    /*
-    Drive forwards until two minerals are centered in the camera's view
-    Speed should be a low value so that the robot doesn't overshoot the minerals
-     */
-    public void driveToMinerals(double maxTimeS){
-
-        int CAMERA_MAX_LEFT = 0;
-        int CAMERA_MAX_RIGHT = 1920; //this should be the same as the horizontal number of pixels of the camera
-        int ERROR_THRESHOLD = 20; //maximum tolerance for the minerals being off center
-
-        boolean centered = false;
-
-        int mineral1XLeft;
-        int mineral2XRight;
-
-        int leftDifference;
-        int rightDifference;
-
-        timer.reset();
-
-        if (opModeIsActive()) {
-            /** Start Tensor Flow Object Detection. */
-            if (tfod != null) {
-                tfod.activate();
-            }
-
-            motorL1.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            motorL2.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            motorR1.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            motorR2.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-            motorL1.setPower(0.3);
-            motorL2.setPower(0.3);
-            motorR1.setPower(0.3);
-            motorR2.setPower(0.3);
-
-            while (opModeIsActive() && timer.seconds() < maxTimeS && !centered) {
-                if (tfod != null) {
-                    List<Recognition> updatedRecognitions = tfod.getUpdatedRecognitions();
-                    if (updatedRecognitions != null) {
-                        telemetry.addData("# Object Detected", updatedRecognitions.size());
-                        if (updatedRecognitions.size() == 2) {
-                            motorL1.setPower(0.1);
-                            motorL2.setPower(0.1);
-                            motorR1.setPower(0.1);
-                            motorR2.setPower(0.1);
-
-                            mineral1XLeft = (int) updatedRecognitions.get(0).getLeft();
-                            mineral2XRight = (int) updatedRecognitions.get(1).getRight();
-
-                            leftDifference = Math.abs(CAMERA_MAX_LEFT - mineral1XLeft);
-                            rightDifference = Math.abs(CAMERA_MAX_RIGHT - mineral2XRight);
-
-                            if(leftDifference > rightDifference &&
-                                    Math.abs(leftDifference - rightDifference) > ERROR_THRESHOLD){
-
-                                motorL1.setPower(0.1); //the minerals are too far right in the camera frame, so keep driving
-                                motorL2.setPower(0.1);
-                                motorR1.setPower(0.1);
-                                motorR2.setPower(0.1);
-                            } else if(leftDifference < rightDifference &&
-                                    Math.abs(leftDifference - rightDifference) > ERROR_THRESHOLD){
-
-                                motorL1.setPower(-0.1); //the minerals are too far left in the camera frame, so drive in reverse
-                                motorL2.setPower(-0.1);
-                                motorR1.setPower(-0.1);
-                                motorR2.setPower(-0.1);
-                            } else {
-                                motorL1.setPower(0); //the minerals are within the threshold, so stop movement
-                                motorL2.setPower(0);
-                                motorR1.setPower(0);
-                                motorR2.setPower(0);
-
-                                centered = true; //breaks loop
-
-                            }
-
-                        }
-                    }
-                }
-            }
         }
     }
